implementacion metodos de RepositorioPlanta

+++ Obligatorio1/Dominio/EntidadesNegocio/FichaCuidados.cs
@@ -8,7 +8,7 @@ namespace Dominio.EntidadesNegocio
     {
         public int Id { get; set; }
         public string TipoIluminacion { get; set; }
-        public double Temperatura { get; set; }
+        public decimal Temperatura { get; set; }
         public string FrecuenciaRiego { get; set; }
 
         public bool SoyValido()
